+++ judo-runtime-core-jsl-itest/pom.xml
@@ -103,6 +103,31 @@
                 </configuration>
             </plugin>
             -->
+
+            <!--plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <version>3.8.0</version>
+                <configuration>
+                    <annotationProcessors>
+                        <annotationProcessor>hu.blackbelt.judo.annotation.RequirementProcessor</annotationProcessor>
+                    </annotationProcessors>
+                    <debug>true</debug>
+                    <optimize>true</optimize>
+                    <compilerArguments>
+                        <AaddGeneratedAnnotation>true</AaddGeneratedAnnotation>
+                        <Adebug>true</Adebug>
+                    </compilerArguments>
+                    <annotationProcessorPaths>
+                        <annotationProcessorPath>
+                            <groupId>hu.blackbelt.judo.annotation</groupId>
+                            <artifactId>judo-annotations</artifactId>
+                            <version>1.0-SNAPSHOT</version>
+                        </annotationProcessorPath>
+                    </annotationProcessorPaths>
+                </configuration>
+
+            </plugin-->
         </plugins>
     </build>
 
@@ -150,6 +175,14 @@
         <dependency>
             <groupId>hu.blackbelt.judo</groupId>
             <artifactId>judo-sdk-common</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>hu.blackbelt.judo.annotation</groupId>
+            <artifactId>judo-annotations</artifactId>
+            <version>1.0-SNAPSHOT</version>
+            <optional>true</optional>
+            <scope>compile</scope>
         </dependency>
     </dependencies>
 </project>
