+++ judo-runtime-core-jsl-itest/src/test/java/hu/blackbelt/judo/runtime/core/jsl/AssociationRelationshipsTest.java
@@ -24,7 +24,7 @@ import com.google.inject.Inject;
 import com.google.inject.Module;
 import hu.blackbelt.judo.runtime.core.jsl.itest.associationrelationships.guice.associationrelationships.AssociationRelationshipsDaoModules;
 import hu.blackbelt.judo.runtime.core.jsl.itest.associationrelationships.sdk.associationrelationships.associationrelationships.*;
-import hu.blackbelt.judo.test.Requirement;
+import hu.blackbelt.judo.annotation.Requirement;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
