+++ judo-runtime-core-jsl-itest/src/test/java/hu/blackbelt/judo/runtime/core/jsl/FiltersTest.java
@@ -27,7 +27,7 @@ import hu.blackbelt.judo.runtime.core.jsl.itest.primitives.guice.primitives.Prim
 import hu.blackbelt.judo.runtime.core.jsl.itest.primitives.sdk.primitives.primitives.MyEntityWithOptionalFields;
 import hu.blackbelt.judo.runtime.core.jsl.itest.primitives.sdk.primitives.primitives.MyEnum;
 import hu.blackbelt.judo.sdk.query.*;
-import hu.blackbelt.judo.test.Requirement;
+import hu.blackbelt.judo.annotation.Requirement;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
