+++ judo-runtime-core-jsl-itest/src/test/java/hu/blackbelt/judo/runtime/core/jsl/TernaryTest.java
@@ -24,7 +24,7 @@ import com.google.inject.Inject;
 import com.google.inject.Module;
 import hu.blackbelt.judo.runtime.core.jsl.itest.ternarytest.guice.ternarytest.TernaryTestDaoModules;
 import hu.blackbelt.judo.runtime.core.jsl.itest.ternarytest.sdk.ternarytest.ternarytest.AAA;
-import hu.blackbelt.judo.test.Requirement;
+import hu.blackbelt.judo.annotation.Requirement;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
